Update RWP_Factions_Town.xml
EMERGENCY HOT FIX.

This solves an error that users who are running both this Mod and the Call of Cthulhu Faction Mod ( https://www.steamcommunity.com/sharedfiles/filedetails/?id=764375772 ) will encounter when trying to generate a new World. The misspelled Pawn Kind Defs in this Mod's Town Faction cause all Worlds to fail to generate and the game to be unable to generate any Pawn Group for the Faction, the latter of which can cause crashes to the desktop.