Very Minor
It's not a slur where I come from unless you call, say, a Korean or Lao the term- no one wants to be confused for a very different race and culture -but it is in California. Although I don't like not using words I like to use, I am changing it in case Mlie or other users are from California.